samples/ow-bigquery-integration-api/requirements.txt
-i https://pypi.org/simple
annotated-types==0.5.0 ; python_version >= '3.7'
anyio==4.0.0rc1 ; python_version >= '3.8'
cachetools==5.3.1 ; python_version >= '3.7'
certifi==2023.7.22 ; python_version >= '3.6'
charset-normalizer==3.2.0 ; python_full_version >= '3.7.0'
click==8.1.6 ; python_version >= '3.7'
exceptiongroup==1.1.2 ; python_version < '3.11'
fastapi==0.99.0
google-api-core[grpc]==2.12.0.dev0 ; python_version >= '3.7'
google-auth==2.22.0 ; python_version >= '3.6'
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3 ; python_version >= '3.7'
google-crc32c==1.5.0 ; python_version >= '3.7'
google-resumable-media==2.5.0 ; python_version >= '3.7'
googleapis-common-protos==1.60.0 ; python_version >= '3.7'
greenlet==3.0.0a1 ; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
grpcio==1.57.0rc1 ; python_version >= '3.7'
grpcio-status==1.57.0rc1
h11==0.14.0 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
mangum==0.17.0
packaging==23.1 ; python_version >= '3.7'
proto-plus==1.22.3 ; python_version >= '3.6'
protobuf==4.24.0rc2 ; python_version >= '3.7'
pyasn1==0.5.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pydantic==1.10.12
pydantic-core==2.4.0 ; python_version >= '3.7'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.31.0 ; python_version >= '3.7'
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
sqlalchemy==2.0.19
starlette==0.27.0
typing-extensions==4.7.1 ; python_version >= '3.7'
urllib3==1.26.16 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
uvicorn==0.23.2
